+++ assigner/commands/open.py
@@ -2,27 +2,29 @@ import logging
 
 from requests.exceptions import HTTPError
 
+from assigner.backends.base import RepoError
+from assigner.backends.decorators import requires_config_and_backend
 from assigner.roster_util import get_filtered_roster
-from assigner.baserepo import Access, Repo, RepoError, StudentRepo
-from assigner.config import config_context
 from assigner import progress
 
 help = "Grants students access to their repos"
 
 logger = logging.getLogger(__name__)
 
-def open_assignment(repo, student):
+
+def open_assignment(repo, student, access):
     try:
         logging.debug("Opening %s...", repo.name)
-        repo.add_member(student["id"], Access.developer)
+        repo.add_member(student["id"], access)
     except HTTPError as e:
         if e.response.status_code == 409:
             logging.warning("%s is already a member of %s.", student["username"], repo.name)
         else:
             raise
 
-@config_context
-def open_all_assignments(conf, args):
+
+@requires_config_and_backend
+def open_all_assignments(conf, backend, args):
     """Adds each student in the roster to their respective homework
     repositories as Developers so they can pull/commit/push their work.
     """
@@ -37,15 +39,15 @@ def open_all_assignments(conf, args):
     for student in progress.iterate(roster):
         username = student["username"]
         student_section = student["section"]
-        full_name = StudentRepo.build_name(semester, student_section,
-                                           hw_name, username)
+        full_name = backend.student_repo.build_name(semester, student_section,
+                                                    hw_name, username)
 
         try:
-            repo = StudentRepo(backend_conf, namespace, full_name)
+            repo = backend.student_repo(backend_conf, namespace, full_name)
             if "id" not in student:
-                student["id"] = Repo.get_user_id(username, backend_conf)
+                student["id"] = backend.repo.get_user_id(username, backend_conf)
 
-            open_assignment(repo, student)
+            open_assignment(repo, student, backend.access.developer)
             count += 1
         except RepoError:
             logging.warning("Could not add %s to %s.", username, full_name)