+++ assigner/commands/status.py
@@ -2,19 +2,19 @@ import logging
 from datetime import datetime
 
 from prettytable import PrettyTable
-from assigner import progress
 
+from assigner import progress
+from assigner.backends.base import RepoError
+from assigner.backends.decorators import requires_config_and_backend
 from assigner.roster_util import get_filtered_roster
-from assigner.config import config_context
-from assigner.baserepo import Access, Repo, StudentRepo, RepoError
 
 help = "Retrieve status of repos"
 
 logger = logging.getLogger(__name__)
 
 
-@config_context
-def status(conf, args):
+@requires_config_and_backend
+def status(conf, backend, args):
     """Retrieves and prints the status of repos"""
     hw_name = args.name
 
@@ -42,12 +42,12 @@ def status(conf, args):
         name = student["name"]
         username = student["username"]
         student_section = student["section"]
-        full_name = StudentRepo.build_name(semester, student_section,
-                                           hw_name, username)
+        full_name = backend.student_repo.build_name(semester, student_section,
+                                                    hw_name, username)
 
         row = [i+1, student_section, username, name, "", "", "", "", ""]
 
-        repo = StudentRepo(backend_conf, namespace, full_name)
+        repo = backend.student_repo(backend_conf, namespace, full_name)
 
         if not repo.already_exists():
             row[4] = "Not Assigned"
@@ -56,7 +56,7 @@ def status(conf, args):
 
         if "id" not in student:
             try:
-                student["id"] = Repo.get_user_id(username, backend_conf)
+                student["id"] = backend.repo.get_user_id(username, backend_conf)
             except RepoError:
                 row[4] = "No Gitlab user"
                 output.add_row(row)
@@ -67,12 +67,13 @@ def status(conf, args):
             row[4] = "Not Opened"
             output.add_row(row)
             continue
-
         if repo.info["archived"]:
-            row[4] = "Archived"
+            row[4] = 'Archived'
         else:
-            level = Access([s["access_level"] for s in members if s["id"] == student["id"]][0])
-            row[4] = "Open" if level is Access.developer else "Locked"
+            level = backend.access(
+                [s["access_level"] for s in members if s["id"] == student["id"]][0]
+            )
+            row[4] = "Open" if level is backend.access.developer else "Locked"
 
         branches = repo.list_branches()