+++ assigner/commands/unlock.py
@@ -1,7 +1,6 @@
 import logging
 
 from assigner import manage_repos
-from assigner.baserepo import Access
 
 help = "Unlock student's repo"
 
@@ -14,7 +13,7 @@ def unlock(args):
     #pylint: disable=no-value-for-parameter
     return manage_repos(
         args,
-        lambda repo, student: repo.edit_member(student["id"], Access.developer)
+        lambda repo, student: repo.unlock(student["id"])
     )